Refactored PHP.cs and fixed a BUG with php serialization float/doubles
