--- ZORGATH/ChamptionOfNewerthRank.cs
@@ -1,27 +1,0 @@
-﻿namespace ZORGATH;
-
-public enum ChampionsOfNewerthRank
-{
-    NoRank,
-    BronzeFive,
-    BronzeFour,
-    BronzeThree,
-    BronzeTwo,
-    BronzeOne,
-    SilverFive,
-    SilverFour,
-    SilverThree,
-    SilverTwo,
-    SilverOne,
-    GoldFour,
-    GoldThree,
-    GoldTwo,
-    GoldOne,
-    DiamondThree,
-    DiamondTwo,
-    DiamondOne,
-    LegendaryTwo,
-    LegendaryOne,
-    Immortal
-}
-
