+++ ZORGATH/GameAPI/PlinkoController.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace ZORGATH.GameAPI;
+
+public class PlinkoController : Controller
+{
+    // GET
+    public IActionResult Index()
+    {
+        return View();
+    }
+}
