+++ ZORGATH/GameAPI/README.md
@@ -1,0 +1,53 @@
+# Folder structure
+
+
+
+### There is 44-46 different cases, those should be grouped into their respective "domain" or "group"
+
+
+```
+add_room
+auth
+get_match_stats
+new_ignored
+nick2id
+pre_auth
+remove_ignored
+srpAuth
+remove_room
+set_rank
+clan_nameCheck
+clan_list
+client_events_info
+get_account_all_hero_stats
+get_hero_usage_graph
+get_hero_usage_list
+get_hero_stats
+get_seasons
+get_special_messages
+autocompleteNicks
+get_upgrades
+selected_upgrades
+reset_stats
+take_mastery_reward
+boost_match_mastery
+show_mastery_stats
+show_mastery_stats
+get_account_mastery
+get_guide_list_filtered
+get_guide
+show_stats
+get_campaign_hero_stats
+grab_last_matches_from_nick
+get_player_award_summ
+change_buddy_group
+get_daily_special
+show_simple_stats
+claim_season_rewards
+server_list
+logout
+remove_buddy2
+delete_notification
+remove_all_notifications
+switch_auth
+```
