+++ ZORGATH/GameAPI/Support/ChampionsOfNewerthRank.cs
@@ -1,0 +1,27 @@
+﻿namespace ZORGATH.GameAPI.Support;
+
+public enum ChampionsOfNewerthRank
+{
+    NoRank,
+    BronzeFive,
+    BronzeFour,
+    BronzeThree,
+    BronzeTwo,
+    BronzeOne,
+    SilverFive,
+    SilverFour,
+    SilverThree,
+    SilverTwo,
+    SilverOne,
+    GoldFour,
+    GoldThree,
+    GoldTwo,
+    GoldOne,
+    DiamondThree,
+    DiamondTwo,
+    DiamondOne,
+    LegendaryTwo,
+    LegendaryOne,
+    Immortal
+}
+
